+++ test/support/variant_post.ex
@@ -46,7 +46,8 @@ defmodule AshStorage.Test.VariantPost do
   use Ash.Resource,
     domain: AshStorage.Test.Domain,
     data_layer: Ash.DataLayer.Ets,
-    extensions: [AshStorage]
+    extensions: [AshStorage],
+    otp_app: :ash_storage
 
   ets do
     private? true